+++ codec-transforms/src/main/java/org/jwcarman/codec/transform/compress/DeflateCodec.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jwcarman.codec.transform;
+package org.jwcarman.codec.transform.compress;
 
 import java.io.IOException;
 import java.io.InputStream;